Adapt integration tests
