+++ tests/mysql_shell/clients/test_instance.py
