+++ tests/mysql_shell/executors/test_local.py
