+++ tests/mysql_shell/helpers.py
@@ -5,10 +5,9 @@ import os
 import threading
 import time
 from contextlib import contextmanager
-from typing import Any
 
 from mysql_shell.executors import LocalExecutor
-from mysql_shell.models import ConnectionDetails, VariableScope
+from mysql_shell.models import ConnectionDetails
 
 TEST_CLUSTER_NAME = "test-cluster"
 TEST_CLUSTER_HOST = "0.0.0.0"
@@ -46,28 +45,3 @@ def temp_process(query: str):
         yield
     finally:
         thread.join()
-
-
-@contextmanager
-def temp_variable(scope: VariableScope, name: str, new_value: Any):
-    """Context manager to run a piece of code with a variable changed."""
-    executor = build_local_executor(
-        username=os.environ.get("MYSQL_USERNAME"),
-        password=os.environ.get("MYSQL_PASSWORD"),
-    )
-
-    set_query = "SET @@{scope}.{name} = {value}"
-    get_query = "SELECT @@{scope}.{name} AS {name}"
-    get_query = get_query.format(scope=scope.value, name=name)
-
-    old_value = executor.execute_sql(get_query)[0][name]
-    old_value = f"'{old_value}'" if isinstance(old_value, str) else old_value
-    new_value = f"'{new_value}'" if isinstance(new_value, str) else new_value
-
-    try:
-        set_query_new = set_query.format(scope=scope.value, name=name, value=new_value)
-        executor.execute_sql(set_query_new)
-        yield
-    finally:
-        set_query_old = set_query.format(scope=scope.value, name=name, value=old_value)
-        executor.execute_sql(set_query_old)
