+++ tests/mysql_shell/models/test_connection.py
