+++ tests/mysql_shell_contrib/__init__.py
