+++ tests/mysql_shell_contrib/builders/authorization/__init__.py
@@ -1,0 +1,2 @@
+# Copyright 2025 Canonical Ltd.
+# See LICENSE file for licensing details.
