+++ tests/mysql_shell_contrib/builders/locking/test_charm.py
@@ -5,8 +5,8 @@ import os
 
 import pytest
 
-from mysql_shell.builders import CharmLockingQueryBuilder
 from mysql_shell.executors import LocalExecutor
+from mysql_shell_contrib.builders import CharmLockingQueryBuilder
 
 from ...helpers import build_local_executor
